CLEANED interface declarations. They can be found in the file interfaces.ts now
