+++ .vscode/settings.json
@@ -1,0 +1,7 @@
+{
+    "workbench.colorCustomizations": {
+        "activityBar.background": "#062F52",
+        "titleBar.activeBackground": "#084172",
+        "titleBar.activeForeground": "#F8FBFE"
+    }
+}
