+++ src/elasticsearch.ts
@@ -1,25 +1,4 @@
-// Defines the shape of a hit.
-interface ShardHit {
-  _id: string;
-  _source: {
-    "http://www w3 org/2000/01/rdf-schema#comment"?: string[];
-    "http://www w3 org/2004/02/skos/core#definition"?: string[];
-  };
-}
-
-// Defines the shape of the fetched object in elasticSuggestions().
-interface ShardResponse {
-  timed_out: boolean;
-  hits: {
-    hits: ShardHit[];
-  };
-}
-
-// Defines the shape of the Elasticsearch recommendations.
-export interface Result {
-  iri: string;
-  description?: string;
-}
+import { Result, ShardHit, ShardResponse } from "./interfaces"
 
 /**
  * Assigns the Elasticsearch query according to the given category.
