+++ test.txt
@@ -1,5 +1,6 @@
 yarn run v1.22.19
-$ node ./dist/recommend.js -t person -c class -e nde
+$ node ./dist/recommend.js -t Person -c class -e nde
+[ 'defaultClass.rq', 'defaultClass.rq', 'defaultClass.rq' ]
 test: class
 
 http://www.cidoc-crm.org/cidoc-crm/E21_Person
@@ -84,4 +85,4 @@ https://schema.org/UnitPriceSpecification
 description: The price asked for a given offer by the respective organization or person.
 score: 1
 
-Done in 1.16s.
+Done in 0.16s.
